refactor : index <-> position 매퍼를 OmokBoardView로 이동
